requirements.txt
python-telegram-bot==21.4
python-dotenv
requests
SQLAlchemy
pandas
aiohttp
numpy